Add standalone samples build support
