+++ .github/workflows/release.yml
@@ -78,7 +78,7 @@ jobs:
       - name: 🧹 Clean Previous Builds
         run: make clean_all
 
-      - name: 🔨 Build Release Artifact
+      - name: 🔨 Build Thunder Release Artifacts
         run: |
           # Define platform matrix as arrays
           PLATFORMS=(
@@ -97,18 +97,42 @@ jobs:
             OS="${platform%%:*}"
             ARCH="${platform#*:}"
             
-            echo "🔨 Building for $OS/$ARCH..."
+            echo "🔨 Building Thunder backend for $OS/$ARCH..."
             make build_backend OS=$OS ARCH=$ARCH
             
             # Optional: Add a small delay to prevent resource contention
             sleep 1
           done
           
-          echo "✅ All platforms built successfully!"
+          echo "✅ Thunder built for all platforms successfully!"
 
-          echo "🏗️ Building Sample applications..."
+      - name: 🔨 Build Sample Release Artifacts
+        run: |
           make build_samples
-          echo "✅ Sample applications built successfully!"
+
+          # Define platform matrix as arrays
+          PLATFORMS=(
+            "win:x64"
+            "linux:x64"
+            "linux:arm64"
+            "macos:x64"
+            "macos:arm64"
+          )
+
+          # Loop through the platform matrix
+          for platform in "${PLATFORMS[@]}"; do
+            # Split the platform string into OS and ARCH
+            OS="${platform%%:*}"
+            ARCH="${platform#*:}"
+            
+            echo "🔨 Building Samples for $OS/$ARCH..."
+            make package_samples OS=$OS ARCH=$ARCH
+            
+            # Optional: Add a small delay to prevent resource contention
+            sleep 1
+          done
+
+          echo "✅ Sample built for all platforms successfully!"
 
       - name: 📝 Read Updated Version
         id: version
